requirements.txt
Authlib==1.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
Flask==2.2.2
Flask-MySQLdb==1.0.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql==0.0.3
mysqlclient==2.1.1
pycparser==2.21
PyJWT==2.4.0
python-dotenv==0.20.0
pytz==2022.2.1
requests==2.28.1
stripe==4.1.0
twilio==7.13.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.8.1
